Edited Specs model weight table. - command makemigrations and migrate; - put password on DB settings;

+++ cars/models.py
@@ -5,4 +5,4 @@ from django.db import models
 class Specs(models.Model):
     name = models.CharField(max_length=20)
     price = models.DecimalField(max_digits=8, decimal_places=2)
-    weight = models.PositiveIntegerField()
+    weight = models.DecimalField(max_digits=7, decimal_places=3)
